refactor(core): move special optimization to file-level sources
