+++ CMakeLists.txt
@@ -92,6 +92,44 @@ target_sources(
   PRIVATE ${${PROJECT_NAME}_MIDDLEWARE_NESTED_SUBDIR_SOURCES}
 )
 
+file(GLOB_RECURSE _xr_force_o3_sources CONFIGURE_DEPENDS
+     "${CMAKE_CURRENT_SOURCE_DIR}/src/*_o3.cpp"
+)
+list(SORT _xr_force_o3_sources)
+
+file(GLOB_RECURSE _xr_force_os_sources CONFIGURE_DEPENDS
+     "${CMAKE_CURRENT_SOURCE_DIR}/src/*_os.cpp"
+)
+list(SORT _xr_force_os_sources)
+
+if(_xr_force_o3_sources)
+  if(CMAKE_CXX_COMPILER_ID STREQUAL "MSVC")
+    set_source_files_properties(${_xr_force_o3_sources} PROPERTIES COMPILE_OPTIONS "/O2")
+  elseif(CMAKE_CXX_COMPILER_ID STREQUAL "GNU" OR
+         CMAKE_CXX_COMPILER_ID STREQUAL "Clang" OR
+         CMAKE_CXX_COMPILER_ID STREQUAL "AppleClang")
+    set_source_files_properties(${_xr_force_o3_sources} PROPERTIES COMPILE_OPTIONS "-O3")
+  else()
+    message(WARNING
+            "LibXR file-level optimized sources are not configured for compiler: "
+            "${CMAKE_CXX_COMPILER_ID}")
+  endif()
+endif()
+
+if(_xr_force_os_sources)
+  if(CMAKE_CXX_COMPILER_ID STREQUAL "MSVC")
+    set_source_files_properties(${_xr_force_os_sources} PROPERTIES COMPILE_OPTIONS "/O1")
+  elseif(CMAKE_CXX_COMPILER_ID STREQUAL "GNU" OR
+         CMAKE_CXX_COMPILER_ID STREQUAL "Clang" OR
+         CMAKE_CXX_COMPILER_ID STREQUAL "AppleClang")
+    set_source_files_properties(${_xr_force_os_sources} PROPERTIES COMPILE_OPTIONS "-Os")
+  else()
+    message(WARNING
+            "LibXR file-level size-optimized sources are not configured for compiler: "
+            "${CMAKE_CXX_COMPILER_ID}")
+  endif()
+endif()
+
 include(${CMAKE_CURRENT_SOURCE_DIR}/src/core/CMakeLists.txt)
 
 # The current tree still relies on basename and short subtree includes across major layers.
