+++ src/core/print/writer/writer_float_runtime.hpp
@@ -105,60 +105,3 @@ bool Writer::ExceedsFixedIntegerDigits(Float value, uint8_t precision)
  * @param ch 待追加字符 / Character to append
  * @return 成功返回 `true`，否则返回 `false` / Returns `true` on success, otherwise `false`
  */
-inline bool Writer::AppendBufferChar(char* buffer, size_t capacity, size_t& size,
-                                     char ch)
-{
-  if (size >= capacity)
-  {
-    return false;
-  }
-  buffer[size++] = ch;
-  return true;
-}
-
-/**
- * @brief 向局部有界浮点格式缓冲区追加一段文本 / Append one text span into a bounded local float-format buffer
- * @param buffer 目标缓冲区 / Destination buffer
- * @param capacity 缓冲区总容量 / Total buffer capacity
- * @param size 当前已保留长度；成功时会推进 / Current retained size; advanced on success
- * @param text 待追加文本片段 / Text span to append
- * @return 成功返回 `true`，否则返回 `false` / Returns `true` on success, otherwise `false`
- */
-inline bool Writer::AppendBufferText(char* buffer, size_t capacity, size_t& size,
-                                     std::string_view text)
-{
-  if (text.size() > capacity - size)
-  {
-    return false;
-  }
-  std::memcpy(buffer + size, text.data(), text.size());
-  size += text.size();
-  return true;
-}
-
-/**
- * @brief 向有界缓冲区追加一个带前导零的 `uint32_t` 十进制片段 / Append one zero-padded `uint32_t` decimal fragment into the bounded buffer
- * @param buffer 目标缓冲区 / Destination buffer
- * @param capacity 缓冲区总容量 / Total buffer capacity
- * @param size 当前已保留长度；成功时会推进 / Current retained size; advanced on success
- * @param value 待追加的无符号值 / Unsigned value to append
- * @param width 目标零填充宽度 / Target zero-padded width
- * @return 成功返回 `true`，否则返回 `false` / Returns `true` on success, otherwise `false`
- */
-inline bool Writer::AppendBufferU32ZeroPad(char* buffer, size_t capacity, size_t& size,
-                                           uint32_t value, uint8_t width)
-{
-  char digits[UnsignedDigitCapacity<uint32_t, 10>()];
-  size_t digit_count = AppendUnsigned<10>(digits, value);
-  size_t zero_count = (width > digit_count) ? static_cast<size_t>(width) - digit_count : 0;
-
-  for (size_t i = 0; i < zero_count; ++i)
-  {
-    if (!AppendBufferChar(buffer, capacity, size, '0'))
-    {
-      return false;
-    }
-  }
-
-  return AppendBufferText(buffer, capacity, size, std::string_view(digits, digit_count));
-}
