+++ src/core/print/writer/writer_reader.hpp
@@ -14,15 +14,13 @@ class Writer::CodeReader
    * @brief 从单个编译字节块起点构造读取器 / Create a reader over the beginning of one compiled byte blob
    * @param codes 指向单个编译字节块起点的指针 / Pointer to the beginning of one compiled byte blob
    */
-  explicit CodeReader(const uint8_t* codes) : pos_(codes), base_(codes)
-  {
-  }
+  explicit CodeReader(const uint8_t* codes);
 
   /**
    * @brief 读取下一条运行期操作码 / Read the next runtime opcode
    * @return 返回下一条解码后的运行期操作码 / Returns the next decoded runtime opcode
    */
-  [[nodiscard]] FormatOp ReadOp() { return static_cast<FormatOp>(*pos_++); }
+  [[nodiscard]] FormatOp ReadOp();
 
   /**
    * @brief 读取编译期发射器按本机字节序写入的 POD 值 / Read a native-endian POD value emitted by the compile-time emitter
@@ -42,7 +40,7 @@ class Writer::CodeReader
    * @brief 读取 `GenericField` 载荷中的语义类型字节 / Read the semantic type byte carried by one `GenericField` payload
    * @return 返回解码后的字段语义类型 / Returns the decoded field semantic type
    */
-  [[nodiscard]] FormatType ReadFormatType() { return static_cast<FormatType>(*pos_++); }
+  [[nodiscard]] FormatType ReadFormatType();
 
   /**
    * @brief 读取紧跟在 `GenericField` 类型字节后的 4 字节字段载荷 / Read the 4-byte field payload that follows one `GenericField` type byte
@@ -53,25 +51,13 @@ class Writer::CodeReader
    * flags、fill、width、precision 这 4 个字节，顺序固定。
    * @return 返回解码后的运行期字段规格 / Returns the decoded runtime field spec
    */
-  [[nodiscard]] Spec ReadSpec()
-  {
-    return Spec{.flags = *pos_++,
-                .fill = static_cast<char>(*pos_++),
-                .width = *pos_++,
-                .precision = *pos_++};
-  }
+  [[nodiscard]] Spec ReadSpec();
 
   /**
    * @brief 读取内嵌在记录流中的短文本 / Read a null-terminated short text payload embedded in the record stream
    * @return 返回该内嵌文本片段的视图 / Returns a view of the embedded text span
    */
-  [[nodiscard]] std::string_view ReadInlineText()
-  {
-    auto text = reinterpret_cast<const char*>(pos_);
-    size_t size = std::strlen(text);
-    pos_ += size + 1;
-    return std::string_view(text, size);
-  }
+  [[nodiscard]] std::string_view ReadInlineText();
 
   /**
    * @brief 读取指向尾部文本池的偏移和长度 / Read an offset-size pair pointing into the trailing text pool
@@ -80,13 +66,7 @@ class Writer::CodeReader
    * 该偏移已经按最终代码块起点完成重定位。
    * @return 返回尾部文本池中被引用的文本片段 / Returns the referenced text span inside the trailing text pool
    */
-  [[nodiscard]] std::string_view ReadTextRef()
-  {
-    auto offset = Read<uint16_t>();
-    auto size = Read<uint16_t>();
-    auto text = reinterpret_cast<const char*>(base_ + offset);
-    return std::string_view(text, size);
-  }
+  [[nodiscard]] std::string_view ReadTextRef();
 
  private:
   const uint8_t* pos_ = nullptr;
@@ -103,7 +83,7 @@ class Writer::ArgumentReader
    * @brief 从单个运行期参数打包字节块构造读取器 / Create a reader over one packed runtime argument blob
    * @param data 指向运行期参数打包字节块的指针 / Pointer to the packed runtime argument blob
    */
-  explicit ArgumentReader(const uint8_t* data) : pos_(data) {}
+  explicit ArgumentReader(const uint8_t* data);
 
   /**
    * @brief 以无对齐要求的方式读取一个已打包参数值 / Read one packed argument value without requiring alignment
